Fixed SELF_TEST_ENABLED being defined
By having SELF_TEST_ENABLED on line 29, the code enters the conditional statement on line 107, which will have it run a modified self-test function. According to line 112, and confirmed by multiple tests, the performSelfTest() function will end in an infinite loop on both success or fail, which binds the code and does not let it continue to the MODE_WAIT_CONFIG runMode.